OD export limited to spacecraft estimations
